Ajout des erreurs
Ajout des erreurs pour les GET


Ajout des erreurs pour le put et post


Ajout des erreurs pour le delete

+++ src/routes/deletePokemon.js
@@ -1,16 +1,24 @@
-const { Pokemon } = require('../db/sequelize')
+const {Pokemon} = require('../db/sequelize')
 
 module.exports = (app) => {
     app.delete('/api/pokemons/:id', (req, res) => {
         Pokemon.findByPk(req.params.id).then(pokemon => {
             const pokemonDeleted = pokemon;
-            Pokemon.destroy({
-                where: { id: pokemon.id }
+            if (pokemon === null) {
+                const message = "le pokémon n'existe pas."
+                return res.status(404).json({message})
+            }
+            return Pokemon.destroy({
+                where: {id: pokemon.id}
             })
                 .then(_ => {
                     const message = `Le pokémon avec l'identifiant n°${pokemonDeleted.id} a bien été supprimé.`
-                    res.json({message, data: pokemonDeleted })
+                    res.json({message, data: pokemonDeleted})
                 })
         })
+            .catch(error => {
+                const message = "le pokémon n'a pas pu etre supprimé";
+                res.status(500).json({message, data: error})
+            })
     })
 }
